Add SolverAdapter and some docstrings
